Fixing the currentBlock function and making the update of the player skin more readable

+++ Render/image_define/Levels/Block.java
@@ -6,7 +6,15 @@ public class Block {
     Rectangle2D block;
     Image skin;
 
-    public Block(long x,long y,long width,long height,String type){
+    /**
+     * Create a block that can be use to define the level
+     * @param x position of the upper left corner
+     * @param y position of the upper left corner
+     * @param width of the block
+     * @param height of the block
+     * @param type to define the texture
+     */
+    public Block(double x,double y,double width,double height,String type){
         block = new Rectangle2D(x,y,width,height);
         switch (type){
             //later grass and more argument will be add
